Score represented as integer (1 pawn = score 100)

+++ MantaChessEngine/Doubles/FakeEvaluator.cs
@@ -8,15 +8,15 @@ namespace MantaChessEngine
 {
     public class FakeEvaluator : IEvaluator
     {
-        private IEnumerator<float> _enumerator;
+        private IEnumerator<int> _enumerator;
 
-        public FakeEvaluator(IEnumerable<float> scores)
+        public FakeEvaluator(IEnumerable<int> scores)
         {
             _enumerator = scores.GetEnumerator();
             EvaluateCalledCounter = 0;
         }
 
-        public float Evaluate(IBoard board)
+        public int Evaluate(IBoard board)
         {
             EvaluateCalledCounter++;
             _enumerator.MoveNext();
